changes ьфву фссщквшצשmade according to comments on Follow model

+++ posts/views.py
@@ -59,7 +59,6 @@ def profile(request, username):
         following_count = None
     context = {
         'user': authorised_user,
-        'authorised_user': authorised_user,
         'page_obj': page_obj,
         'posts': posts,
         'author': author,
@@ -90,56 +89,53 @@ def post_detail(request, post_id):
     }
     return render(request, template, context)
 
-
+@login_required()
 def post_create(request):
     authorised_user = request.user
-    if request.user.is_authenticated:
-        if request.method == 'POST':
-            form = PostForm(request.POST, files=request.FILES or None)
-            if form.is_valid():
-                try:
-                    new_post = form.save(commit=False)
-                    new_post.author_id = request.user.id
-                    new_post.save()
-                    return redirect('posts:profile', username=request.user.username)
-                except Exception as e:
-                    form.add_error(None, 'Ошибка добавления поста')
-        else:
-            form = PostForm()
-        template = 'posts/post_create.html'
-        context = {
-            'authorised_user': authorised_user,
-            'form': form,
-            'title': 'Новый пост',
-        }
-        return render(request, template, context)
-    else:
-        template = 'users/login.html'
-        return render(request, template)
-
-
-def post_edit(request, post_id):
-    authorised_user = request.user
-    post = get_object_or_404(Post, id=post_id)
     if request.method == 'POST':
-        form = PostForm(request.POST or None, files=request.FILES or None, instance=post)
+        form = PostForm(request.POST, files=request.FILES or None)
         if form.is_valid():
             try:
-                form.save()
+                new_post = form.save(commit=False)
+                new_post.author_id = request.user.id
+                new_post.save()
                 return redirect('posts:profile', username=request.user.username)
             except Exception as e:
                 form.add_error(None, 'Ошибка добавления поста')
     else:
-        form = PostForm(instance=post)
-
-    template = 'posts/post_edit.html'
+        form = PostForm()
+    template = 'posts/post_create.html'
     context = {
         'authorised_user': authorised_user,
-        'post': post,
         'form': form,
-        'is_edit': True
+        'title': 'Новый пост',
     }
     return render(request, template, context)
+
+@login_required()
+def post_edit(request, post_id):
+    authorised_user = request.user
+    post = get_object_or_404(Post, id=post_id)
+    if authorised_user.id == post.author.id:
+        if request.method == 'POST':
+            form = PostForm(request.POST or None, files=request.FILES or None, instance=post)
+            if form.is_valid():
+                try:
+                    form.save()
+                    return redirect('posts:profile', username=request.user.username)
+                except Exception as e:
+                    form.add_error(None, 'Ошибка добавления поста')
+        else:
+            form = PostForm(instance=post)
+
+        template = 'posts/post_edit.html'
+        context = {
+            'authorised_user': authorised_user,
+            'post': post,
+            'form': form,
+            'is_edit': True
+        }
+        return render(request, template, context)
 
 
 @login_required
@@ -179,8 +175,7 @@ def follow_index(request):
     context = {
         'authorised_user': authorised_user,
         'page_obj': page_obj,
-        'paginator': paginator,
-        # 'following': following,
+        'paginator': paginator
     }
     return render(request, 'posts/follow_index.html', context)
 
@@ -190,8 +185,8 @@ def profile_follow(request, username):
     authorised_user = request.user
     author = get_object_or_404(User, username=username)
     follow = Follow.objects.filter(author=author, user=authorised_user)
-    if author != request.user:
-        Follow.objects.get_or_create(user=authorised_user, author=author)
+    if len(follow) == 0 and author != request.user:
+        follow = Follow.objects.get_or_create(user=authorised_user, author=author)
     context = {
         'authorised_user': authorised_user,
         'author': author,
@@ -204,7 +199,8 @@ def profile_unfollow(request, username):
     # Дизлайк, отписка
     authorised_user = request.user
     author = get_object_or_404(User, username=username)
-    Follow.objects.filter(user=authorised_user, author=author).delete()
+    if len(Follow.objects.filter(user=authorised_user, author=author)) > 0:
+        Follow.objects.filter(user=authorised_user, author=author).delete()
     context = {
         'authorised_user': authorised_user,
         'author': author,
